Rename to CommonAnalyzers and AnalyzerBuilder
Drop `asciiFolderWithLower` from CommonAnalyzers
