+++ lucene/src/main/scala/textmogrify/lucene/CommonAnalyzers.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2022 Pig.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package textmogrify.lucene
+
+import org.apache.lucene.analysis.Analyzer
+import org.apache.lucene.analysis.en.EnglishAnalyzer
+
+object CommonAnalyzers {
+
+  def englishStandard(): Analyzer = new EnglishAnalyzer()
+
+  def porterStemmer(): Analyzer =
+    AnalyzerBuilder.default.withLowerCasing.withPorterStemmer.build
+
+  def asciiFolder(): Analyzer =
+    AnalyzerBuilder.default.withASCIIFolding.build
+
+}
